pyproject.toml
[tool.poetry]
name = "ai-final-project"
version = "0.1.0"
description = ""
authors = ["Andres Rubio <andres.rubio@artificial-solutions.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.10.18"
llama-index-readers-web = "^0.1.7"
install = "^1.3.5"
langchain-community = "^0.0.28"
langchain-core = "^0.1.32"
faiss-cpu = "^1.8.0"
llama-index-vector-stores-faiss = "^0.1.2"
python-dotenv = "1.0.0"
llama-index-embeddings-instructor = "^0.1.3"
chainlit = "^1.0.401"
llama-index-callbacks-wandb = "^0.1.2"
streamlit = "^1.32.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
